customer controller endpoints

+++ src/main/java/com/eqriesracingteam/garage/repository/CarRepository.java
@@ -3,5 +3,6 @@ package com.eqriesracingteam.garage.repository;
 import com.eqriesracingteam.garage.model.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
+
 }
